Added 3 Default bookshelves
Added the 3 default bookshelves, Read, Want to Read, and Reading. Users can now add books to that to keep track of their books while reusing our shelf functionality. Added a reset bookshelves to default button so you can get the 3 bookshelves.Added a progress bar in the book detail page under the reading status tab. Added a way for users to overwrite the page count since the API counts a lot of the end pages that arent apart of the book